requirements/base.txt
# Core FastAPI dependencies
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.9
pydantic[email]==2.11.9

# Database dependencies
sqlalchemy[asyncio]==2.0.43
aiomysql==0.2.0
alembic==1.14.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP and networking
httpx==0.27.0
aiohttp==3.10.11

# Redis for caching
redis[hiredis]==5.1.1

# Environment and configuration
python-dotenv==1.0.1

# Logging and monitoring
structlog==24.4.0

# Utilities
email-validator==2.2.0
python-slugify==8.0.4
pytz==2024.2

# Image processing
pillow==11.0.0

# WebSocket support (included in FastAPI)
websockets==13.1

# Data validation and serialization
pydantic-settings==2.6.1